Organized Code and Improved Evolution Algorithm
All the value functions have a separate file. All the stop conditions have a separate file. The code for the data structures has a separate file.

The evolution algorithm is based on removing or inserting buildings, rather than replacing. It also uses weighted randomness instead of pure randomness. (Weights can still be set to equal in the function parameters)
