+++ README.md
@@ -1,2 +1,12 @@
 # CCC
 Cookie Clicker Calculator
+
+Attempts to find the optimal play style in the game Cookie Clicker.
+Specifically, for speedrunning the 'hardcore' style in the fastest amount of time.
+
+Rules are:
+  1. Get to 1 billion total earned cookies in the shortest amount of time.
+        The 1 billion figure includes cookies that have already been spent, as
+        long as they were earned at one point.
+  2. You may not buy upgrades to any buildings.
+  3. You may not collect golden cookies.
